Move Badge class to preferences package
